Fixed conversion warning.
git-svn-id: https://svn.code.sf.net/p/half/code/trunk@210 ba856b29-48f5-46b1-a3e5-b459205bce89

+++ include/half.hpp
@@ -399,7 +399,7 @@ namespace half_float
 				24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 13 };
 			uint32 bits;// = *reinterpret_cast<uint32*>(&value);		//violating strict aliasing!
 			std::memcpy(&bits, &value, sizeof(float));
-			uint16 hbits = base_table[bits>>23] + ((bits&0x7FFFFF)>>shift_table[bits>>23]);
+			uint16 hbits = base_table[bits>>23] + static_cast<uint16>((bits&0x7FFFFF)>>shift_table[bits>>23]);
 			if(R == std::round_to_nearest)
 				hbits += (((bits&0x7FFFFF)>>(shift_table[bits>>23]-1))|(((bits>>23)&0xFF)==102)) & 
 					/*(((((static_cast<uint32>(1)<<(shift_table[bits>>23]-1))-1)&bits)!=0)|hbits) &*/ ((hbits&0x7C00)!=0x7C00);
